Require payment of HNY for BYOT Gardens. Update use of OG token vs GDN token for BYOT Gardens. Make Honey deposit a min required.

+++ packages/hardhat/contracts/external/IIncentivisedPriceOracleFactory.sol
@@ -5,7 +5,7 @@ contract IIncentivisedPriceOracleFactory {
     // The returned address can be sent some _incentiveToken as an incentive for users to update the price
     function newIncentivisedPriceOracle(
         address _honeyswapFactory,
-        uint _windowSize,
+        uint256 _windowSize,
         uint8 _granularity,
         address _incentiveToken,
         uint256 _percentIncentivePerCall,
